Time: 331 ms (5.02%), Space: 111.2 MB (5.55%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -26,7 +26,7 @@ public:
             int high=i;
             int pointer=-1;
             while(low <= high){
-                long long int mid=(low+high)/2; // target value
+                int mid=(low+high)/2; // target value
                 long long int window=i-mid+1;// 13 is included
 
                 long long int exp_sum=window*nums[i];
